renamed delete function folder name
