+++ .github/workflows/webscraping-pipeline.yml
@@ -4,7 +4,7 @@ on:
         - main
       paths:
         - WebScraping/functions/parser/**
-        - WebScraping/functions/deleteDB/**
+        - WebScraping/functions/deleteTable/**
         - WebScraping/functions/insertData/**
 jobs:
     build-deploy:
