+++ WebScraping/functions/deleteTable/lambda_function.py
